部分代码逻辑修改

+++ pojo/src/main/java/com/hundsun/fund/transaction/vo/UserInfoVO.java
@@ -12,9 +12,12 @@ import java.math.BigDecimal;
 @Data
 public class UserInfoVO implements Serializable {
 
+    // 用户名
     private String name;
 
+    // 余额
     private BigDecimal balance;
 
-    private Byte status;
+    // 状态
+    private Integer status;
 }
